% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFinal.R
\name{executeIncidencePrevalenceFinal}
\alias{executeIncidencePrevalenceFinal}
\title{Execute the CHAPTER study}
\usage{
executeIncidencePrevalenceFinal(
  dbConnection,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  writePrefix,
  outputFolder,
  databaseId,
  readCohorts = TRUE
)
}
\arguments{
\item{dbConnection}{An object of type \code{dbConnection} as created using the
\code{\link[DBI]{dbConnect}} function in the
DBI package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write privileges in this schema.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{databaseId}{A short string for identifying the database (e.g.
'Synpuf').}

\item{readCohorts}{A boolean to choose whether to read and instantiate the initial cohorts
from the available jsons.}

\item{writePrefix.}{Write prefix for the tables to be created in the write schema.}

\item{createCovidCohorts}{A boolean to choose whether to create the COVID related cohorts.}
}
\description{
Execute the CHAPTER study
}
\details{
This function executes the incidence analyses of CHAPTER study.
}
